+++ kie-wb-common-dmn/kie-wb-common-dmn-webapp-kogito-runtime/src/main/resources/org/kie/workbench/common/dmn/showcase/DMNKogitoRuntimeWebapp.gwt.xml
@@ -19,6 +19,7 @@
 <module>
 
   <inherits name="org.kie.workbench.common.dmn.webapp.kogito.common.DMNWebappKogitoCommon"/>
+  <inherits name="org.kie.workbench.common.stunner.kogito.runtime.StunnerKogitoRuntimeCommon"/>
 
   <!-- We only need Chrome -->
   <set-property name="user.agent" value="safari"/>
