+++ kie-wb-common-dmn/kie-wb-common-dmn-webapp-kogito-testing/pom.xml
@@ -696,6 +696,7 @@
                 <compileSourcesArtifact>org.uberfire:uberfire-wires-core-grids</compileSourcesArtifact>
 
                 <compileSourcesArtifact>org.uberfire:appformer-js-bridge</compileSourcesArtifact>
+                <compileSourcesArtifact>org.uberfire:appformer-kogito-api</compileSourcesArtifact>
 
               </compileSourcesArtifacts>
             </configuration>
