+++ kie-wb-common-stunner/kie-wb-common-stunner-core/kie-wb-common-stunner-commons/kie-wb-common-stunner-client-common/src/main/java/org/kie/workbench/common/stunner/core/client/command/RedoCommandHandler.java
@@ -19,8 +19,8 @@ import javax.enterprise.context.Dependent;
 import javax.enterprise.event.Event;
 import javax.inject.Inject;
 
-import org.appformer.kogito.bridge.client.stateControl.registry.CommandRegistry;
-import org.appformer.kogito.bridge.client.stateControl.registry.DefaultCommandRegistry;
+import org.appformer.kogito.api.stateControl.registry.CommandRegistry;
+import org.appformer.kogito.api.stateControl.registry.DefaultCommandRegistry;
 import org.kie.workbench.common.stunner.core.client.canvas.event.registration.RegisterChangedEvent;
 import org.kie.workbench.common.stunner.core.client.session.ClientSession;
 import org.kie.workbench.common.stunner.core.command.Command;
