+++ kie-wb-common-stunner/kie-wb-common-stunner-core/kie-wb-common-stunner-commons/kie-wb-common-stunner-client-common/src/main/resources/org/kie/workbench/common/stunner/core/StunnerClientCommon.gwt.xml
@@ -25,6 +25,7 @@
   <inherits name="org.kie.workbench.common.stunner.core.StunnerBackendApi"/>
   <inherits name="org.kie.workbench.common.stunner.core.StunnerCoreCommon"/>
   <inherits name="org.uberfire.client.views.PatternFly"/>
+  <inherits name="org.appformer.kogito.AppformerKogitoAPI"/>
 
   <source path="client"/>
   <source path="" includes="StunnerCoreEntryPoint.java"/>
