+++ kie-wb-common-stunner/kie-wb-common-stunner-extensions/kie-wb-common-stunner-kogito/kie-wb-common-stunner-kogito-client/pom.xml
@@ -90,11 +90,6 @@
       <groupId>com.google.elemental2</groupId>
       <artifactId>elemental2-promise</artifactId>
     </dependency>
-    <dependency>
-      <groupId>org.uberfire</groupId>
-      <artifactId>appformer-kogito-bridge</artifactId>
-    </dependency>
-
 
     <!-- Stunner -->
     <dependency>
