+++ kie-wb-common-stunner/kie-wb-common-stunner-extensions/kie-wb-common-stunner-kogito/kie-wb-common-stunner-kogito-client/src/main/resources/org/kie/workbench/common/stunner/kogito/StunnerKogitoClient.gwt.xml
@@ -38,7 +38,6 @@
 
   <!-- Kogito -->
   <inherits name="org.kie.workbench.common.kogito.KogitoClient"/>
-  <inherits name="org.appformer.kogito.bridge.AppformerKogitoBridge"/>
 
   <source path="client"/>
