+++ kie-wb-common-stunner/kie-wb-common-stunner-extensions/kie-wb-common-stunner-kogito/kie-wb-common-stunner-kogito-runtime-common/src/test/java/org/kie/workbench/common/stunner/kogito/runtime/client/session/command/impl/KogitoUndoSessionCommandTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.kie.workbench.common.stunner.kogito.client.session.command.impl;
+package org.kie.workbench.common.stunner.kogito.runtime.client.session.command.impl;
 
 import org.appformer.kogito.bridge.client.stateControl.interop.StateControl;
 import org.junit.Before;
@@ -23,32 +23,25 @@ import org.junit.runner.RunWith;
 import org.kie.workbench.common.stunner.core.client.canvas.AbstractCanvas;
 import org.kie.workbench.common.stunner.core.client.canvas.AbstractCanvasHandler;
 import org.kie.workbench.common.stunner.core.client.canvas.controls.keyboard.KeyboardControl;
-import org.kie.workbench.common.stunner.core.client.command.CanvasViolation;
-import org.kie.workbench.common.stunner.core.client.command.RedoCommandHandler;
 import org.kie.workbench.common.stunner.core.client.command.SessionCommandManager;
 import org.kie.workbench.common.stunner.core.client.session.ClientSession;
 import org.kie.workbench.common.stunner.core.client.session.impl.EditorSession;
-import org.kie.workbench.common.stunner.core.command.Command;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 
 import static org.mockito.Matchers.any;
-import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
-public class KogitoRedoSessionCommandTest {
+public class KogitoUndoSessionCommandTest {
 
     @Mock
     private StateControl stateControl;
 
     @Mock
     private SessionCommandManager<AbstractCanvasHandler> sessionCommandManager;
-
-    @Mock
-    private RedoCommandHandler<Command<AbstractCanvasHandler, CanvasViolation>> redoCommandHandler;
 
     @Mock
     protected KeyboardControl<AbstractCanvas, ClientSession> keyboardControl;
@@ -58,35 +51,31 @@ public class KogitoRedoSessionCommandTest {
 
     private boolean envelopeAvailable = false;
 
-    private KogitoRedoSessionCommand redoSessionCommand;
+    private KogitoUndoSessionCommand undoSessionCommand;
 
     @Before
     public void setup() {
         when(session.getKeyboardControl()).thenReturn(keyboardControl);
 
-        redoSessionCommand = new KogitoRedoSessionCommand(sessionCommandManager, redoCommandHandler, () -> envelopeAvailable, () -> stateControl);
+        undoSessionCommand = new KogitoUndoSessionCommand(sessionCommandManager, () -> envelopeAvailable, () -> stateControl);
     }
 
     @Test
     public void testBindCommandInKogito() {
         this.envelopeAvailable = true;
 
-        redoSessionCommand.bind(session);
-
-        verify(redoCommandHandler).setSession(eq(session));
+        undoSessionCommand.bind(session);
 
         verify(keyboardControl, never()).addKeyShortcutCallback(any());
-        verify(stateControl).setRedoCommand(any());
+        verify(stateControl).setUndoCommand(any());
     }
 
     @Test
     public void testBindCommandOutsideKogito() {
 
-        redoSessionCommand.bind(session);
-
-        verify(redoCommandHandler).setSession(eq(session));
+        undoSessionCommand.bind(session);
 
         verify(keyboardControl).addKeyShortcutCallback(any());
-        verify(stateControl, never()).setRedoCommand(any());
+        verify(stateControl, never()).setUndoCommand(any());
     }
 }
