+++ kie-wb-common-stunner/kie-wb-common-stunner-sets/kie-wb-common-stunner-bpmn/kie-wb-common-stunner-bpmn-kogito-runtime/pom.xml
@@ -620,8 +620,18 @@
       <artifactId>kie-wb-common-stunner-kogito-client</artifactId>
       <scope>provided</scope>
     </dependency>
+    <dependency>
+      <groupId>org.kie.workbench.stunner</groupId>
+      <artifactId>kie-wb-common-stunner-kogito-runtime-common</artifactId>
+      <scope>provided</scope>
+    </dependency>
 
-    <!--  Appformer Kogito Bridge -->
+    <!--  Appformer Kogito -->
+    <dependency>
+      <groupId>org.uberfire</groupId>
+      <artifactId>appformer-kogito-api</artifactId>
+      <scope>provided</scope>
+    </dependency>
     <dependency>
       <groupId>org.uberfire</groupId>
       <artifactId>appformer-kogito-bridge</artifactId>
@@ -789,11 +799,13 @@
             <compileSourcesArtifact>org.kie.workbench.stunner:kie-wb-common-stunner-forms-client</compileSourcesArtifact>
             <compileSourcesArtifact>org.kie.workbench.stunner:kie-wb-common-stunner-kogito-api</compileSourcesArtifact>
             <compileSourcesArtifact>org.kie.workbench.stunner:kie-wb-common-stunner-kogito-client</compileSourcesArtifact>
+            <compileSourcesArtifact>org.kie.workbench.stunner:kie-wb-common-stunner-kogito-runtime-common</compileSourcesArtifact>
 
             <!-- Appformer JS Bridge -->
             <compileSourcesArtifact>org.uberfire:appformer-js-bridge</compileSourcesArtifact>
 
             <!-- Appformer Kogito Bridge -->
+            <compileSourcesArtifact>org.uberfire:appformer-kogito-api</compileSourcesArtifact>
             <compileSourcesArtifact>org.uberfire:appformer-kogito-bridge</compileSourcesArtifact>
 
           </compileSourcesArtifacts>
