+++ kie-wb-common-stunner/kie-wb-common-stunner-sets/kie-wb-common-stunner-bpmn/kie-wb-common-stunner-bpmn-kogito-runtime/src/main/resources/org/kie/workbench/common/stunner/kogito/KogitoBPMNEditor.gwt.xml
@@ -44,6 +44,7 @@
   <!-- Stunner -->
   <inherits name="org.kie.workbench.common.stunner.client.StunnerWidgets"/>
   <inherits name="org.kie.workbench.common.stunner.kogito.StunnerKogitoClient"/>
+  <inherits name="org.kie.workbench.common.stunner.kogito.runtime.StunnerKogitoRuntimeCommon"/>
 
   <!-- Form modeler -->
   <inherits name="org.kie.workbench.common.stunner.forms.StunnerFormsClient"/>
